Extract AppButton animations into useButtonAnimation hook
