+++ src/frontend/components/common/AppButton.tsx
@@ -1,4 +1,4 @@
-import React, { useRef, useEffect, useState } from 'react';
+import React from 'react';
 import {
     Text,
     StyleProp,
@@ -6,13 +6,12 @@ import {
     TouchableOpacity,
     ViewStyle,
     Animated,
-    Easing,
     ActivityIndicator,
 } from 'react-native';
 import { colors } from '../../theme/colors';
 import { typography } from '../../theme/typography';
 import { shadows } from '../../theme/shadows';
-import { duration } from '../../theme/animations';
+import { useButtonAnimation } from '../../hooks/useButtonAnimation';
 
 interface Props {
     title?: string;
@@ -24,101 +23,22 @@ interface Props {
     minLoadTime?: number;
 }
 
-export const AppButton = ({ 
-    title, 
-    onPress, 
-    style, 
-    disabled, 
-    shakeCount, 
+export const AppButton = ({
+    title,
+    onPress,
+    style,
+    disabled,
+    shakeCount,
     isLoading,
-    minLoadTime = 1000, 
+    minLoadTime = 1000,
 }: Props) => {
-    const scaleAnim = useRef(new Animated.Value(1)).current;
-    const shakeAnim = useRef(new Animated.Value(0)).current;
-
-    const [isVisuallyLoading, setIsVisuallyLoading] = useState(false);
-    const loadingStartTime = useRef<number | null>(null);
-    const timeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-    const prevShakeCount = useRef(shakeCount);
-    const pendingShake = useRef(false);
-
-    const triggerShake = () => {
-        Animated.sequence([
-            Animated.timing(shakeAnim, { toValue: 4, duration: 40, useNativeDriver: true }),
-            Animated.timing(shakeAnim, { toValue: -4, duration: 40, useNativeDriver: true }),
-            Animated.timing(shakeAnim, { toValue: 3, duration: 40, useNativeDriver: true }),
-            Animated.timing(shakeAnim, { toValue: -3, duration: 40, useNativeDriver: true }),
-            Animated.timing(shakeAnim, { toValue: 0, duration: 40, useNativeDriver: true }),
-        ]).start();
-    };
-
-    useEffect(() => {
-        if (isLoading) {
-            setIsVisuallyLoading(true);
-            loadingStartTime.current = Date.now();
-            if (timeoutRef.current) clearTimeout(timeoutRef.current);
-        } else if (isVisuallyLoading && loadingStartTime.current !== null) {
-            const timeElapsed = Date.now() - loadingStartTime.current;
-            const timeRemaining = minLoadTime - timeElapsed;
-
-            if (timeRemaining > 0) {
-                timeoutRef.current = setTimeout(() => {
-                    setIsVisuallyLoading(false);
-                    loadingStartTime.current = null;
-                }, timeRemaining);
-            } else {
-                setIsVisuallyLoading(false);
-                loadingStartTime.current = null;
-            }
-        }
-    }, [isLoading, minLoadTime]);
-
-    useEffect(() => {
-        return () => {
-            if (timeoutRef.current) clearTimeout(timeoutRef.current);
-        };
-    }, []);
-
-    useEffect(() => {
-        if (shakeCount === prevShakeCount.current) return;
-        prevShakeCount.current = shakeCount;
-
-        if (!shakeCount) return;
-
-        if (isVisuallyLoading) {
-            pendingShake.current = true;
-        } else {
-            triggerShake();
-        }
-    }, [shakeCount, isVisuallyLoading]);
-
-    useEffect(() => {
-        if (!isVisuallyLoading && pendingShake.current) {
-            pendingShake.current = false;
-            triggerShake();
-        }
-    }, [isVisuallyLoading]);
-
-    const ease = Easing.bezier(0.25, 1, 0.5, 1);
-
-    const handlePressIn = () => {
-        Animated.timing(scaleAnim, {
-            toValue: 0.96,
-            duration: duration.micro,
-            easing: ease,
-            useNativeDriver: true,
-        }).start();
-    };
-
-    const handlePressOut = () => {
-        Animated.timing(scaleAnim, {
-            toValue: 1,
-            duration: duration.micro,
-            easing: ease,
-            useNativeDriver: true,
-        }).start();
-    };
+    const {
+        scaleAnim,
+        shakeAnim,
+        isVisuallyLoading,
+        handlePressIn,
+        handlePressOut,
+    } = useButtonAnimation({ isLoading, shakeCount, minLoadTime });
 
     const isDisabled = disabled || isVisuallyLoading;
