+++ src/frontend/hooks/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from './useButtonAnimation';
